requirements.txt
typer==0.12.5
rich==13.7.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.32
alembic==1.13.2
psycopg[binary]==3.2.1
opencv-python==4.10.0.84
Pillow>=11.1.0
selenium==4.23.1
webdriver-manager==4.0.2
httpx==0.27.2
supabase==2.7.0
gotrue==2.8.0
storage3==0.7.4
postgrest==0.16.9
openai==1.50.2
requests==2.32.3
pillow-heif>=1.1.0
imageio-ffmpeg==0.4.9
